pyproject.toml
[project]
name = "recipe-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain-community>=0.3.20",
    "langchain>=0.3.21",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentence-transformers>=3.4.1",
    "streamlit>=1.43.2",
]
